✨ Feat: 모달 닫기 버튼 공용화 후 적용
